requirements.txt
streamlit==0.68.0
pandas==1.1.3
numpy==1.19.2
pickleshare==0.7.5
Pillow==7.2.0
scikit-learn==0.23.2